Create MatrixChainMultiplication.cpp
Problem Statement: 

Matrix Chain Multiplication | Partition DP Starts

In the coming articles, we will discuss problems related to a new pattern called “Partition DP”.Before proceeding further, let us understand how to identify whether a problem can be solved using this pattern.


To summarize:

Represent the entire array by two indexes i and j. In this question i =1 and j = n. We need to find f(i,j).
Maintain a mini variable to get the minimum answer.
Set a for loop to find the answer( variable k) from i to j-1,
In every iteration find the answer, with the formula discussed above and compare it with the mini value.
Return mini as the answer.